Use the solr:slim docker image
The previous image, `solr:7-alpine` has not been updated in almost a
year. If we believe our components that use Solr are not tied to Solr 7,
we should allow the Solr version to advance. With the solr_version
parameter, the orb's consumers can control which version they wish to
use. The README has been updated to reflect this.