Makes bundle cache key creation more repeatable
In CircleCI, different parallel runs could order the files found for the
bundle cache key differently, resulting in cache misses, even though the
files found and the MD5s were the same. Now, the checksums are sorted.

The file that becomes part of the cache key is also added to the run's
artifacts, so future debugging won't have to do so much guesswork.