Changes the bundle cache key construction
The cache key now is of the form
`v<cache_version>-ruby<ruby_version>-bundler<bundler_version>-bundle<bundle_checksum>`
to allow bundle checksum misses to hit a better, if less specific,
cache. In general, I expect this order to be least-often-changed to
most-often-changed.

Also shows the BUNDLE_CACHE_KEY contents in a step rather than saving
the file as an artifact.